fix regression: restore balanced nodes check for ,
The EXPECT_END state contained ',' which did not check the depth.

reproduce:
echo '"test",' | json2tsv

This wrote nodes[-1].index++ and was treated as valid.
a primitive followed by , was also valid (incorrect), for example:
"true,"

EXPECT_END is "}]," All these states are checked now for the depth <= 0 so
remove the EXPECT_NOTHING state.